fix vault buffer message display

+++ eagleproject/notify/triggers/vault_buffer.py
@@ -28,7 +28,7 @@ def run_trigger(new_logs):
                 "Vault Buffer Updated   🤏",
                 "OUSD Vault buffer was changed to {}%".format(
                     # Always whole numbers?
-                    int(Decimal(buffer_percent_bigint) / Decimal(1e18))
+                    Decimal(buffer_percent_bigint) / Decimal(1e18)
                 )
             )
         )
